Bug 1481225: Avoid confusing console spam on native app shutdown. r=aswan
Differential Revision: https://phabricator.services.mozilla.com/D3540

UltraBlame original commit: 9e05198aa74aa76790001aa403d6d7e33b49f9dd

+++ toolkit/components/extensions/NativeMessaging.jsm
@@ -251,7 +251,9 @@ var NativeApp = class extends EventEmitter {
       
       
       let timer = setTimeout(() => {
-        this.proc.kill(GRACEFUL_SHUTDOWN_TIME);
+        if (this.proc) {
+          this.proc.kill(GRACEFUL_SHUTDOWN_TIME);
+        }
       }, GRACEFUL_SHUTDOWN_TIME);
 
       let promise = Promise.all([
